requirements.txt
gunicorn
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
Werkzeug
pyjwt
